# Numerical operation optimization and overflow handling