```ruby
a=0.1
while a < 100000
  a+=1
  a -=0.1
  a *= 1.001
  a /= 1.001
end
```
| Method      | Mean     | Error     | StdDev    | Allocated |
|------------ |---------:|----------:|----------:|----------:|
| MRubyCS     | 5.893 ms | 0.4677 ms | 0.3094 ms |     178 B |
| MRubyCS(PR) | 3.084 ms | 0.0158 ms | 0.0105 ms |     176 B |
| MRubyNative | 4.415 ms | 0.0419 ms | 0.0277 ms |       3 B |