@hadashiA 2025-06-28
Thanks a lot!

Just a question:
Is converting to arithId (0,1,2,3) once faster than branching with enum(OpCode.Add,Sub,Mul,Div) ?