@Akeit0 2025-06-28
Oh.
I had tried it myself and found this one to be faster, but it seemed to be just a margin of error, so I put it back.